# [Merged by Bors] - perf: no redundant deduplication in `Equiv.Perm.cycleFactorsFinset`