@Komyyy 2025-05-31
@eric-wieser
As a result of the benchmark, the performance improvement is somewhat limited, but since it does have an effect, I merge this branch.

bors r+